nmr|tests: Move build test to integration tests
