+++ packages/nmr/vitest.integration.config.ts
@@ -1,0 +1,1 @@
+export { default } from '../../config/vitest.integration.config.ts';
