+++ packages/nmr/vitest.standalone.config.ts
@@ -1,0 +1,1 @@
+export { default } from '../../config/vitest.standalone.config.ts';
